- Add new user level error code. - Update authenticate and authorize middleware. - Removed redundency from auth context by stroing only user value in the context. - Removed claims and roles from context. - Minor fixes.
