+++ apps/backend/business/web/errors/codes.go
@@ -9,6 +9,8 @@ const (
 	NotFound       ErrorCode = "RESOURCE_NOT_FOUND"
 	AlreadyExists  ErrorCode = "RESOURCE_ALREADY_EXISTS"
 
+	UserNotVerified ErrorCode = "USER_NOT_VERIFIED"
+
 	BadRequest          ErrorCode = "BAD_REQUEST"
 	InternalServerError ErrorCode = "INTERNAL_ERROR"
 	ServiceUnavailable  ErrorCode = "SERVICE_UNAVAILABLE"
